// src/api/generated/model/userRoleUpdateDtoRole.ts
/**
 * Generated by orval v7.20.0 🍺
 * Do not edit manually.
 * Cosy API
 * Management API for Cosy (Cost Optimised Server Yard).
 * OpenAPI spec version: v1.0
 */

export type UserRoleUpdateDtoRole = typeof UserRoleUpdateDtoRole[keyof typeof UserRoleUpdateDtoRole];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UserRoleUpdateDtoRole = {
  OWNER: 'OWNER',
  ADMIN: 'ADMIN',
  QUOTA_USER: 'QUOTA_USER',
} as const;